+++ src/main/scala/io/delta/sql/DeltaSparkSessionExtension.scala
@@ -16,185 +16,19 @@
 
 package io.delta.sql
 
-import scala.collection.JavaConverters._
+import io.delta.sql.parser.DeltaSqlParser
 
-import io.delta.sql.parser.DeltaSqlBaseParser._
-import io.delta.sql.parser._
-import io.delta.tables.execution.VacuumTableCommand
-import org.antlr.v4.runtime._
-import org.antlr.v4.runtime.atn.PredictionMode
-import org.antlr.v4.runtime.misc.{Interval, ParseCancellationException}
-import org.antlr.v4.runtime.tree._
+import org.apache.spark.sql.SparkSessionExtensions
 
-import org.apache.spark.sql.{AnalysisException, SparkSessionExtensions}
-import org.apache.spark.sql.catalyst.expressions.Expression
-import org.apache.spark.sql.catalyst.{FunctionIdentifier, TableIdentifier}
-import org.apache.spark.sql.catalyst.parser.{ParseErrorListener, ParseException, ParserInterface}
-import org.apache.spark.sql.catalyst.parser.ParserUtils.{string, withOrigin}
-import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
-import org.apache.spark.sql.catalyst.trees.Origin
-import org.apache.spark.sql.types.{DataType, StructType}
-
+/**
+ * An extension for Spark SQL to support Delta's SQL commands. This one can be enabled by setting
+ * `spark.sql.extensions` to `io.delta.sql.DeltaSparkSessionExtension` before creating your
+ * `SparkSession`.
+ */
 class DeltaSparkSessionExtension extends (SparkSessionExtensions => Unit) {
   override def apply(extensions: SparkSessionExtensions): Unit = {
     extensions.injectParser { (session, parser) =>
       new DeltaSqlParser(parser)
     }
-  }
-}
-
-/** Fork from `org.apache.spark.sql.catalyst.parser.UpperCaseCharStream` */
-class UpperCaseCharStream(wrapped: CodePointCharStream) extends CharStream {
-  override def consume(): Unit = wrapped.consume
-  override def getSourceName(): String = wrapped.getSourceName
-  override def index(): Int = wrapped.index
-  override def mark(): Int = wrapped.mark
-  override def release(marker: Int): Unit = wrapped.release(marker)
-  override def seek(where: Int): Unit = wrapped.seek(where)
-  override def size(): Int = wrapped.size
-
-  override def getText(interval: Interval): String = {
-    // ANTLR 4.7's CodePointCharStream implementations have bugs when
-    // getText() is called with an empty stream, or intervals where
-    // the start > end. See
-    // https://github.com/antlr/antlr4/commit/ac9f7530 for one fix
-    // that is not yet in a released ANTLR artifact.
-    if (size() > 0 && (interval.b - interval.a >= 0)) {
-      wrapped.getText(interval)
-    } else {
-      ""
-    }
-  }
-
-  override def LA(i: Int): Int = {
-    val la = wrapped.LA(i)
-    if (la == 0 || la == IntStream.EOF) la
-    else Character.toUpperCase(la)
-  }
-}
-
-/**
- * A SQL parser that tries to parse Delta commands. If failng to parse the SQL text, it will
- * forward the call to `delegate`.
- */
-class DeltaSqlParser(val delegate: ParserInterface) extends ParserInterface {
-  private val builder = new DeltaSqlAstBuilder
-
-  override def parsePlan(sqlText: String): LogicalPlan = parse(sqlText) { parser =>
-    builder.visit(parser.singleStatement()) match {
-      case plan: LogicalPlan => plan
-      case x => delegate.parsePlan(sqlText)
-    }
-  }
-
-  protected def parse[T](command: String)(toResult: DeltaSqlBaseParser => T): T = {
-    val lexer = new DeltaSqlBaseLexer(
-      new UpperCaseCharStream(CharStreams.fromString(command)))
-    lexer.removeErrorListeners()
-    lexer.addErrorListener(ParseErrorListener)
-
-    val tokenStream = new CommonTokenStream(lexer)
-    val parser = new DeltaSqlBaseParser(tokenStream)
-    parser.addParseListener(PostProcessor)
-    parser.removeErrorListeners()
-    parser.addErrorListener(ParseErrorListener)
-
-    try {
-      try {
-        // first, try parsing with potentially faster SLL mode
-        parser.getInterpreter.setPredictionMode(PredictionMode.SLL)
-        toResult(parser)
-      } catch {
-        case e: ParseCancellationException =>
-          e.printStackTrace()
-
-          // if we fail, parse with LL mode
-          tokenStream.seek(0) // rewind input stream
-          parser.reset()
-
-          // Try Again.
-          parser.getInterpreter.setPredictionMode(PredictionMode.LL)
-          toResult(parser)
-      }
-    } catch {
-      case e: ParseException if e.command.isDefined =>
-        throw e
-      case e: ParseException =>
-        throw e.withCommand(command)
-      case e: AnalysisException =>
-        val position = Origin(e.line, e.startPosition)
-        throw new ParseException(Option(command), e.message, position, position)
-    }
-  }
-
-  override def parseExpression(sqlText: String): Expression = delegate.parseExpression(sqlText)
-
-  override def parseTableIdentifier(sqlText: String): TableIdentifier =
-    delegate.parseTableIdentifier(sqlText)
-
-  override def parseFunctionIdentifier(sqlText: String): FunctionIdentifier =
-    delegate.parseFunctionIdentifier(sqlText)
-
-  override def parseTableSchema(sqlText: String): StructType = delegate.parseTableSchema(sqlText)
-
-  override def parseDataType(sqlText: String): DataType = delegate.parseDataType(sqlText)
-}
-
-class DeltaSqlAstBuilder extends DeltaSqlBaseBaseVisitor[AnyRef] {
-
-  override def visitVacuumTable(ctx: VacuumTableContext): AnyRef = withOrigin(ctx) {
-    VacuumTableCommand(
-      Option(ctx.path).map(string),
-      Option(ctx.table).map(visitTableIdentifier),
-      Option(ctx.number).map(_.getText.toDouble),
-      ctx.RUN != null)
-  }
-
-  override def visitSingleStatement(ctx: SingleStatementContext): LogicalPlan = withOrigin(ctx) {
-    visit(ctx.statement).asInstanceOf[LogicalPlan]
-  }
-
-  protected def visitTableIdentifier(ctx: QualifiedNameContext): TableIdentifier = withOrigin(ctx) {
-    ctx.identifier.asScala match {
-      case Seq(tbl) => TableIdentifier(tbl.getText)
-      case Seq(db, tbl) => TableIdentifier(tbl.getText, Some(db.getText))
-      case _ => throw new ParseException(s"Illegal table name ${ctx.getText}", ctx)
-    }
-  }
-
-  override def visitPassThrough(ctx: PassThroughContext): LogicalPlan = null
-}
-
-/** Fork from `org.apache.spark.sql.catalyst.parser.PostProcessor`. */
-case object PostProcessor extends DeltaSqlBaseBaseListener {
-
-  /** Remove the back ticks from an Identifier. */
-  override def exitQuotedIdentifier(ctx: QuotedIdentifierContext): Unit = {
-    replaceTokenByIdentifier(ctx, 1) { token =>
-      // Remove the double back ticks in the string.
-      token.setText(token.getText.replace("``", "`"))
-      token
-    }
-  }
-
-  /** Treat non-reserved keywords as Identifiers. */
-  override def exitNonReserved(ctx: NonReservedContext): Unit = {
-    replaceTokenByIdentifier(ctx, 0)(identity)
-  }
-
-  private def replaceTokenByIdentifier(
-    ctx: ParserRuleContext,
-    stripMargins: Int)(
-    f: CommonToken => CommonToken = identity): Unit = {
-    val parent = ctx.getParent
-    parent.removeLastChild()
-    val token = ctx.getChild(0).getPayload.asInstanceOf[Token]
-    val newToken = new CommonToken(
-      new org.antlr.v4.runtime.misc.Pair(token.getTokenSource, token.getInputStream),
-      DeltaSqlBaseParser.IDENTIFIER,
-      token.getChannel,
-      token.getStartIndex + stripMargins,
-      token.getStopIndex - stripMargins)
-    parent.addChild(new TerminalNodeImpl(f(newToken)))
   }
 }
